Refactor and add new features. I wrote code.
1. Separation of fuzzer state into multiple types
- Settings
- Stats for metadata about the fuzzer execution
- World for effects and coeffects
All of which are contained in Info. Containing that state in a new type
was necessary to make signal handling reliable, don't ask me why.
Something about generic metadata access or reference cycles or race
conditions or memory ownership or whatever. I'll revisit that decision
later, I can't debug this issue anymore, it broke me.

2. Signal handling à la SwiftPM. The way I used to do it was not
reliable, at all.

3. Command line arguments handling. I use SwiftPM’s ArgumentParser for
that.

4. Creation of FuzzerJobsManager to prepare for the “minimize” feature
implementation. But also as a proof of concept for more complex “parent”
processes like, say, whole Mac apps.

5. Creation of ReadGraph for debugging purposes

6. Remove support for programs with more than 2^21 PCs. Because I think
it makes sense. I wasn't able to instrument those programs properly
anyway.

7. Use faster clock() implementation thanks to DispatchTime. It makes a
huge difference

8. Don't use binary search for weightedPick of mutators

9. Speaking of mutators: move from function-based to enum-based mutators
and force weighted mutators to be static (instead of being dependent on
the value that was going to be mutated). For performance reason and also
because I think it's simpler.

10. Experimental: rewrite Feature so that it has a stride of 4 bytes
instead of 9. The performance gain seems to be minimal (~3%) so I'll
probably revert that decision eventually.