Rename “Unit” to “Input” everywhere
This is a large change and doesn't have many benefits, but I really do
find Input clearer than Unit. The name “Unit” is already used in
“Unit Tests”, so it is confusing. And what does it mean anyway?

So I am using Input because it is the **input** of the test function.

But what's really happening here? Well, I'm kind of tired and renaming
random things makes me feel productive.