Add a complexity budget to mutators
I am not sure it is the right thing to do, but I had trouble finding
a solution for my generic array generator. The question is:
If I want to mutate a value of type [T] by appending an element T to it, how do I construct that new element? I can’t create a “random” value of type T because T’s size might be unbounded and I can’t create a “basic” value of T (e.g. an empty array) because then I always choose the same value. So my solution is to give each mutator a complexity budget, which will give a maximum bound on the size of T.